full context refactoring, changed approachto theme changing, some styles fixed
