+++ src/components/signout/Signout.js
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { signOut } from "firebase/auth";
 import { auth } from "../../Firebase";
-import DdMenuContext from "../../context/DdMenuContext";
+import { DdMenuContext } from "context/Context";
 import { ReactComponent as LogoutIcon } from "../../assets/logout.svg";
 import CustomButton from "../customButton/CustomButton";
