+++ src/components/user/User.js
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect } from "react";
-import UserContext from "../../context/UserContext";
+import { UserContext } from "context/Context";
 import "./User.scss";
 
 const User = () => {
