+++ src/components/usersData/UsersData.scss
@@ -41,7 +41,7 @@
             p {
                 font-size: 1vw;
                 text-align: left;
-                color: var(--dark-text-color);
+                color: #1a1a1a;
                 @media screen and (max-width: 767px) {
                     font-size: 2.2vw;
                 }
